magento-engcom/community-portal#240 Redesign Project Details Page
