+++ frontend/src/routes/index.tsx
@@ -5,7 +5,7 @@ import Auth from '../components/Auth';
 import HeadBar from '../components/HeadBar';
 import Profile from '../components/Profile';
 import ProjectGrid from '../components/ProjectGrid';
-import ProjectView from '../components/ProjectView/ProjectView';
+import ProjectDetails from '../components/ProjectDetails';
 
 const Routes = () => (
   <div>
@@ -14,12 +14,14 @@ const Routes = () => (
     <Route exact path="/auth" component={Auth} />
     <Route exact path="/profile" component={Profile}/>
     <Route exact path="/bookmarked" render={
-      props => <ProjectGrid filter="bookmarkedProjects" {...props} />
+      (props: any) => <ProjectGrid filter="bookmarkedProjects" {...props} />
     } />
     <Route exact path="/pledged" render={
-      props => <ProjectGrid filter="pledgedProjects" {...props}/>
+      (props: any) => <ProjectGrid filter="pledgedProjects" {...props} />
     } />
-    <Route exact path="/project/:project_id" component={ProjectView} />
+    <Route exact path="/project/:project_id" component={
+      (props: any) => <ProjectDetails project_id={props.match.params.project_id} />
+    } />
   </div>
 );
